Выполненное заданИсправлены замечания

+++ src/Converter.java
@@ -1,11 +1,11 @@
 public class Converter {
 
     double oneStepLen = 0.75;
-    int oneStepCal = 50;
+    double oneStepCal = 50.0;
     public double distance(int sumSteps){
         return (sumSteps * oneStepLen)/1000;
     }
-    public int ccalories(int sumSteps){
+    public double ccalories(int sumSteps){
         return (sumSteps * oneStepCal)/1000;
     }
 }
